ml_models/skin_lesion_classifier/requirements.txt
# Skin Lesion Classifier Dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=9.0.0
pydantic>=2.0.0

# For image processing
# opencv-python>=4.5.0
# scikit-image>=0.19.0
